feat(auth): start JWT authentification
